Fix character error and typo
Ran code through validator, it found errors in lines 19-24, I'm guessing
due to the way the text was copied and pasted, it did something funky
with the fractions in the measurements. I removed the pasted fractions
and retyped, fixed the problem. Found a missing space in line 9 before
"width". I fixed.